GCP Projects/day_08/ex_07/requirements.txt
google-cloud-bigquery==1.28.0
google-cloud-datastore==1.15.3